# webapi/docs/swagger.yaml
basePath: /
definitions:
  http.CategoryDto:
    properties:
      name:
        type: string
    type: object
  http.WordDtoIn:
    properties:
      category_id:
        type: string
      confirmed_user_id:
        type: string
      difficulty:
        type: string
      eng_text:
        type: string
      native_text:
        type: string
      transcription:
        type: string
    type: object
  http.WordDtoUpdate:
    properties:
      category_id:
        type: string
      difficulty:
        type: string
      eng_text:
        type: string
      native_text:
        type: string
      transcription:
        type: string
    type: object
  models.Category:
    properties:
      id:
        description: потом поправлю
        type: string
      name:
        type: string
    type: object
  models.Word:
    properties:
      category_id:
        type: string
      confirmed_at:
        type: string
      confirmed_status:
        type: string
      confirmed_user_id:
        type: string
      difficulty:
        type: string
      eng_text:
        type: string
      id:
        type: string
      native_text:
        type: string
      transcription:
        type: string
    type: object
info:
  contact: {}
  description: 'objects service: CRUD words and categories'
  title: objects-service
  version: "1.0"
paths:
  /api/categories/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Get all categories
      tags:
      - category
    post:
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/http.CategoryDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security:
      - ApiKeyAuth: []
      summary: Create new category
      tags:
      - category
  /api/categories/{id}:
    delete:
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security:
      - ApiKeyAuth: []
      summary: Delete category by id
      tags:
      - category
    get:
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Get category by id
      tags:
      - category
    put:
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/http.CategoryDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security:
      - ApiKeyAuth: []
      summary: Update category by id
      tags:
      - category
  /api/categories/{id}/words:
    get:
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Word'
            type: array
      summary: Get all words by category id
      tags:
      - category
  /api/words/:
    get:
      parameters:
      - description: Filter by difficulty
        in: query
        name: difficulty
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Word'
            type: array
      summary: Get all words
      tags:
      - word
    post:
      parameters:
      - description: word
        in: body
        name: word
        required: true
        schema:
          $ref: '#/definitions/http.WordDtoIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security:
      - ApiKeyAuth: []
      summary: Create new word
      tags:
      - word
  /api/words/{id}:
    delete:
      parameters:
      - description: word id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security:
      - ApiKeyAuth: []
      summary: Delete word by id
      tags:
      - word
    get:
      parameters:
      - description: word id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Word'
      summary: Get word by id
      tags:
      - word
    put:
      parameters:
      - description: word id
        in: path
        name: id
        required: true
        type: string
      - description: word
        in: body
        name: word
        required: true
        schema:
          $ref: '#/definitions/http.WordDtoUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Word'
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
      security:
      - ApiKeyAuth: []
      summary: Update word by id
      tags:
      - word
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"